// oasis/src/texas/types.gen.ts
export type ErrorResponse = {
  error: OAuthErrorCode;
  error_description: string;
};

/**
 * Identity providers for use with token fetch, exchange and introspection.
 */
export type IdentityProvider =
  | "azuread"
  | "tokenx"
  | "maskinporten"
  | "idporten";
// ...
export type OAuthErrorCode =
  | "invalid_request"
  | "invalid_client"
  | "invalid_grant"
  | "unauthorized_client"
  | "unsupported_grant_type"
  | "invalid_scope"
  | "server_error";

/**
 * Use this data type to exchange a user token for a machine token.
 */
export type TokenExchangeRequest = {
  identity_provider: IdentityProvider;
  /**
   * Force renewal of token. Defaults to false if omitted.
   */
  skip_cache?: boolean | null;
  /**
   * Scope or identifier for the target application.
   */
  target: string;
  /**
   * The user's access token, usually found in the _Authorization_ header in requests to your application.
   */
  user_token: string;
};
// ...
export type TokenResponse = {
  access_token: string;
  /**
   * Token expiry in seconds. Useful for caching purposes.
   */
  expires_in: number;
  token_type: TokenType;
};
// ...
export type TokenExchangeResponses = {
  /**
   * Success
   */
  200: TokenResponse;
};

export type TokenExchangeResponse =
  TokenExchangeResponses[keyof TokenExchangeResponses];